requirements.txt
fastapi
uvicorn
pyyaml
pydantic
pytest
httpx 